requirements.txt
flask
scikit-learn
pandas
numpy
mysql-connector-python

